Replaced Player function bodys to Player.Legacy file
Ist is useless to create "new" legacy files for the same class.
